Moved the bluetoothclient and holder to the data package. The DTOs moved to the model package.
