+++ app/src/main/java/se/ju/student/robomow/model/CollisionAvoidanceCircle.kt
@@ -1,4 +1,4 @@
-package se.ju.student.robomow.data
+package se.ju.student.robomow.model
 
 data class CollisionAvoidanceCircle(
     val x: Float,
