+++ app/src/main/java/se/ju/student/robomow/model/CollisionAvoidanceImage.kt
@@ -1,4 +1,4 @@
-package se.ju.student.robomow.data
+package se.ju.student.robomow.model
 
 data class CollisionAvoidanceImage(
     val imageURL: String
