+++ app/src/main/java/se/ju/student/robomow/ui/MainActivity.kt
@@ -17,8 +17,8 @@ import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import io.reactivex.rxjava3.disposables.Disposable
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
-import se.ju.student.robomow.BluetoothClient
-import se.ju.student.robomow.BluetoothClientHolder
+import se.ju.student.robomow.data.BluetoothClient
+import se.ju.student.robomow.data.BluetoothClientHolder
 import se.ju.student.robomow.R
 import se.ju.student.robomow.api.RoboMowApi
 import javax.inject.Inject
